moved all of graph canvas UI integration into a per document view class added a priority based bus for populating the application main menu from multiple sources including subviews moved all docked windows from the views menu to the tools menu
