+++ Gems/Atom/Tools/AtomToolsFramework/Code/Include/AtomToolsFramework/Document/AtomToolsDocumentMainWindow.h
@@ -32,7 +32,7 @@ namespace AtomToolsFramework
 
         using Base = AtomToolsMainWindow;
 
-        AtomToolsDocumentMainWindow(const AZ::Crc32& toolId, QWidget* parent = 0);
+        AtomToolsDocumentMainWindow(const AZ::Crc32& toolId, const QString& objectName, QWidget* parent = 0);
         ~AtomToolsDocumentMainWindow();
 
         //! Helper function to get the absolute path for a document represented by the document ID
@@ -40,6 +40,9 @@ namespace AtomToolsFramework
 
         //! Retrieves the document ID from a tab with the index
         AZ::Uuid GetDocumentTabId(const int tabIndex) const;
+
+        //! Retrieves the document ID from the currently selected tab
+        AZ::Uuid GetCurrentDocumentId() const;
 
         //! Searches for the tab index corresponding to the document ID, returning -1 if not found
         int GetDocumentTabIndex(const AZ::Uuid& documentId) const;
@@ -77,8 +80,11 @@ namespace AtomToolsFramework
         //! Prompts the user for a selection of documents to open
         virtual AZStd::vector<AZStd::string> GetOpenDocumentParams();
 
+        // AtomToolsMainWindowRequestBus::Handler overrides...
+        void CreateMenus(QMenuBar* menuBar) override;
+        void UpdateMenus(QMenuBar* menuBar) override;
+
     protected:
-        void AddDocumentMenus();
         void AddDocumentTabBar();
 
         void AddRecentFilePath(const AZStd::string& absolutePath);
